feat(api): add booking controller and http file

+++ BookingApp.Application/Requests/Members/Commands/DeleteMember/DeleteMemberCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using BookingApp.Core.Abstractions;
 using BookingApp.Core.Domain.Members.Repositories;
+using BookingApp.Core.Exceptions;
 using MediatR;
 
 namespace BookingApp.Application.Requests.Members.Commands.DeleteMember;
@@ -13,7 +14,13 @@ public class DeleteMemberCommandHandler(
         DeleteMemberCommand request, 
         CancellationToken cancellationToken = default)
     {
-        var member = await membersRepository.GetByIdAsync(request.MemberId, cancellationToken);
+        var member = await membersRepository.GetByIdAsync(
+            request.MemberId, 
+            cancellationToken);
+        
+        if (member is null)
+            throw new NotFoundException("Given member was not found.");
+        
         membersRepository.Remove(member);
         await unitOfWork.SaveChangesAsync(cancellationToken);
     }
